+++ server/src/encrypting.c
@@ -2,10 +2,10 @@
 
 void Preparing_numeral_keys(int sock_fd, SSI client){
 
-	int a1 = rand() / 0b11111111111111111111111111111111;
-	int a2 = rand() / 0b11111111111111111111111111111111;
-	int A1 = Speed_degree_with_mod(PUBLIC_KEY_G, a1, PUBLIC_KEY_P);
-	int A2 = Speed_degree_with_mod(PUBLIC_KEY_G, a2, PUBLIC_KEY_P);
+	size_t a1 = rand() / 0b11111111111111111111111111111111;
+	size_t a2 = rand() / 0b11111111111111111111111111111111;
+	size_t A1 = Speed_degree_with_mod(PUBLIC_KEY_G, a1, PUBLIC_KEY_P);
+	size_t A2 = Speed_degree_with_mod(PUBLIC_KEY_G, a2, PUBLIC_KEY_P);
 
 	char client_message[MAX_COMMAND_LENGHT] = {0};
 	char greetings[MAX_COMMAND_LENGHT] = {0};
@@ -25,10 +25,10 @@ void Preparing_numeral_keys(int sock_fd, SSI client){
 	char * place_B1 = strchr(client_message + 1, ':');
 	char * place_B2 = strchr(place_B1 + 1,':');
 
-	int B1 = atoi(place_B1 + 1);
-	int B2 = atoi(place_B2 + 1);
-	int K1 = Speed_degree_with_mod(PUBLIC_KEY_G, B1, PUBLIC_KEY_P);
-	int K2 = Speed_degree_with_mod(PUBLIC_KEY_G, B2, PUBLIC_KEY_P);
+	size_t B1 = atoi(place_B1 + 1);
+	size_t B2 = atoi(place_B2 + 1);
+	size_t K1 = Speed_degree_with_mod(PUBLIC_KEY_G, B1, PUBLIC_KEY_P);
+	size_t K2 = Speed_degree_with_mod(PUBLIC_KEY_G, B2, PUBLIC_KEY_P);
 
 	printf("KEYS:\n%d\n%d\n", K1, K2);
 }
@@ -41,13 +41,13 @@ void Decoding(){
 
 }
 
-int Speed_degree_with_mod(int g, int x, int p){ // return = g^x % p
+size_t Speed_degree_with_mod(size_t g, size_t x, size_t p){ // return = g^x % p
 	if (x == 0)
 		return 1;
 
 	if (x % 2 != 0)
 		return (Speed_degree_with_mod(g, x - 1, p) * g) % p;
 
-	int y = Speed_degree_with_mod(g, x / 2, p);
+	size_t y = Speed_degree_with_mod(g, x / 2, p);
 	return (y * y) % p;
 }
